requirements.txt
PyDriller==2.3
python-dotenv==1.0.0
groq==0.11.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy<2.0
fastapi==0.115.0
uvicorn==0.30.6
